rename `devices` as `sources`
