--- docs/api/devices.md
@@ -1,11 +1,0 @@
-# devices
-
-Interfaces to external devices like miniscopes and DAQs
-
-## OpalKelly
-
-```{eval-rst}
-.. autoclass:: miniscope_io.devices.opalkelly.okDev
-    :members:
-    :undoc-members:
-```
