+++ docs/api/sources.md
@@ -1,0 +1,11 @@
+# sources
+
+Interfaces to external source devices like miniscopes and DAQs
+
+## OpalKelly
+
+```{eval-rst}
+.. autoclass:: miniscope_io.sources.opalkelly.okDev
+    :members:
+    :undoc-members:
+```
