+++ docs/index.md
@@ -22,13 +22,13 @@ device/update_controller
 :maxdepth: 1
 :caption: API:
 
-api/devices
 api/formats/index
 api/io
 api/logging
 api/models/index
 api/exceptions
 api/plots/index
+api/sources
 api/utils
 api/stream_daq
 api/bit_operation
