+++ miniscope_io/__init__.py
@@ -11,7 +11,7 @@ from miniscope_io.models.config import Config
 BASE_DIR = Path(__file__).parent.resolve()
 DATA_DIR = BASE_DIR / "data"
 CONFIG_DIR = DATA_DIR / "config"
-DEVICE_DIR = BASE_DIR / "devices"
+DEVICE_DIR = BASE_DIR / "sources"
 
 __all__ = [
     "BASE_DIR",
