+++ miniscope_io/cameras/__init__.py
@@ -1,0 +1,3 @@
+"""
+Camera classes :)
+"""
