+++ miniscope_io/cameras/camera.py
@@ -1,0 +1,17 @@
+"""
+ABCs for Camera and Miniscope source classes
+"""
+
+from miniscope_io.models import Source
+
+
+class Camera(Source):
+    """
+    A data source that captures images, often in a sequence known as a "video"
+    """
+
+
+class Miniscope(Camera):
+    """
+    A miniature microscope
+    """
