--- miniscope_io/devices/__init__.py
@@ -1,3 +1,0 @@
-"""
-Control interfaces for external hardware devices
-"""
